Add table-based versions for counting
There is a version with one counter per list (but using
a consumer on a per-item topic), and another version with
one counter per list partition.

I should do a separate per-list table version, but fed from
a consumer of a topic keyed by the list itself. This would
be guaranteed to work as well as the per-partition version,
and would still support multiple concurrent consumers (the
parallelism would be per-list instead of per-item).