Add re-keying version of the consumer
This is the final version of the consumer, which takes advantage of both Kafka and a database table.
It is the easiest to understand, once you realize you can use your own Kafka topics to help with the
processing. This re-keying is similar to what you might do with Kafka Streams.

The commit also updates the README to explain the utility of this sample project, and how students
can try it out themselves if they want to understand the challenges around stateful event processing.