Add example tests for our two good consumers
The per-partition table one is easy, as we can just trigger the consumer ourselves.

The re-keying one is tricky as it needs the actual Kafka server to communicate the two parts of the consumer.
The test setup method has to clear not only the database table, but also the topics we use for the consumer.